J'ai effectuer quelques ajustements dans le but de reussir a bon aissance les test unitaire , j'ai ajoute Instantio dans le Pom.xml , modifier le type de prix de BigDecimal a Double pour les chiffres frottante, adapté les methodes dans le CommandeControllerTest , creer des methodes static sur le TestDataFactory pour les tester sur la classe de test unitaire

+++ commande_service/src/main/java/com/kitabe/commande_service/domaine/CommandeItemsEntite.java
@@ -19,7 +19,7 @@ public class CommandeItemsEntite {
     private String code;
 
    @Column(nullable = false)
-   private BigDecimal prix;
+   private Double prix;
 
    @Column(nullable = false)
    private Integer quantite;
@@ -52,11 +52,11 @@ public class CommandeItemsEntite {
         this.code = code;
     }
 
-    public BigDecimal getPrix() {
+    public Double getPrix() {
         return prix;
     }
 
-    public void setPrix(BigDecimal prix) {
+    public void setPrix(Double prix) {
         this.prix = prix;
     }
